Move delete buttons into their own tsx files
